refactor(cloud-account): mutualize credential helpers, split commands, address review
- Extract BuildCredentialRaw / AddAllRawFlags into the credentials package
  so cy credential and cy cloud-account share one source of truth for the
  per-type switch and the inline-credential flag set.
- Add cyargs.AddUpdateFlag / GetUpdate / UpdateFlag and cyargs.CloudAccountFlag
  constant; make AddCloudProviderFlag and AddColorFlag return the flag name
  for symmetry with the rest of cyargs.
- Split cloud-account list.go into list.go / get.go / delete.go; drop the
  retro-compat --cloud-account flag from get and delete (positional args only).
- Fix shadowed err in cloud-account create (the with_credentials branch was
  silently dropping API errors); explicit error checks throughout.
- Document why GetString/GetBool err returns are intentionally ignored on the
  PATCH-style merge in cloud-account update.
- Make the org-env CRUD middleware test cleanup unlink before delete and use
  a randomized canonical so a panicked run does not pin the next one.
- Pin client/version to v6.10.12-rc (backend tag where meta-gov-env merged).
