+++ client/version
@@ -1,1 +1,1 @@
-0
+v6.10.12-rc
